+++ src/main/java/com/gitlab/cdagaming/craftpresence/handler/gui/GUIHandler.java
@@ -27,10 +27,10 @@ import java.util.List;
 
 public class GUIHandler {
     public boolean openConfigGUI = false;
+    public List<String> GUI_NAMES = new ArrayList<>();
     private String CURRENT_GUI_NAME;
     private Class CURRENT_GUI_CLASS;
     private GuiScreen CURRENT_SCREEN;
-    public List<String> GUI_NAMES = new ArrayList<>();
     private List<Class> GUI_CLASSES = new ArrayList<>();
 
     private static void drawTexturedModalRect(int x, int y, int u, int v, int width, int height, float zLevel) {